requirements.txt
pandas==1.5.3
Pillow==9.5.0
plotly==5.14.0
requests==2.28.2
streamlit==1.20.0
streamlit_lottie==0.0.3
